refactor: Remove the async module
