+++ src/stream.rs
@@ -6,9 +6,10 @@ use std::io::{Bytes, Read};
 #[cfg(feature = "std")]
 use easy::Error;
 
+use Parser;
+
 use primitives::{ConsumedResult, ParseError, StreamError, StringStreamError, UnexpectedParse,
                  CHAR_BOUNDARY_ERROR_MESSAGE};
-
 use primitives::FastResult::*;
 
 pub trait Positioned: StreamOnce {
@@ -718,6 +719,29 @@ impl PointerOffset {
     {
         self.0 -= initial_string as *const T as *const () as usize;
         self.0
+    }
+}
+
+pub fn decode<P>(
+    mut parser: P,
+    src: P::Input,
+    partial_state: &mut P::PartialState,
+) -> Result<(Option<P::Output>, usize), <P::Input as StreamOnce>::Error>
+where
+    P: Parser,
+    P::Input: RangeStream,
+{
+    let start = src.checkpoint();
+    let mut input = src;
+    match parser.parse_with_state(&mut input, partial_state) {
+        Ok(message) => Ok((Some(message), input.distance(&start))),
+        Err(err) => {
+            return if input.is_partial() && err.is_unexpected_end_of_input() {
+                Ok((None, input.distance(&start)))
+            } else {
+                Err(err)
+            }
+        }
     }
 }
